% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age}
\alias{age}
\title{age dataset}
\format{
\subsection{\code{age}}{

A tibble with 48 rows and 3 columns:
\describe{
\item{age}{The age categories of students who took either the ACT or SAT in the sat.act dataset.}
\item{freq}{The frequency of participants who are the given age.}
\item{prop}{The proportion of participants who are the given age.}
}
}
}
\source{
insert source here (url's can be in < >)
}
\usage{
age
}
\description{
A frequency table summarizing the sat.act dataset in the psych package.
}
\keyword{datasets}
