% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.dare.R
\name{AIC.dare}
\alias{AIC.dare}
\title{AIC}
\usage{
\method{AIC}{dare}(object, k = 2)
}
\arguments{
\item{object}{object of class dare}
}
\description{
Akaike Information criteria for dare objects.  Lower is better.
}
