% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA_overlay.R
\name{ANOVA_overlay}
\alias{ANOVA_overlay}
\title{Splitting your data to construct overlays that mimic the four sections of
data used to calculate a one-way ANOVA.}
\usage{
ANOVA_overlay(dv, iv, data, common = mean)
}
\arguments{
\item{dv}{Character: What is the name of your dependent variable?}

\item{iv}{Character: What is the name of your independent variable?}

\item{data}{tibble or data frame: Your dataset containing your variables.}

\item{common}{function: which function should represent the common in the
splitting algorithm?}
}
\value{
a list of four tibbles, representing the four overlays: the data,
the common, the condition, and the residuals.
}
\description{
Splitting your data to construct overlays that mimic the four sections of
data used to calculate a one-way ANOVA.
}
\examples{
\dontrun{
# using tapply to see the data split by condition, assuming a mean common
tapply(mtcars$wt,mtcars$am,mean)

# using ANOVA_overlay
ANOVA_overlay("wt", "am", mtcars)
ANOVA_overlay("wt", "am", mtcars, common = median)
ANOVA_overlay("wt", "gear", mtcars)
}
}
