% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesToG.R
\name{bayesToG}
\alias{bayesToG}
\title{Convert a Bayes Factor to a g-prior}
\usage{
bayesToG(bayesFactor)
}
\arguments{
\item{bayesFactor}{the Bayes factor}
}
\value{
The g-prior associated with the inputted Bayes factor
}
\description{
Convert a Bayes Factor to a g-prior
}
