% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.dare.R
\name{BIC.dare}
\alias{BIC.dare}
\title{BIC}
\usage{
\method{BIC}{dare}(object)
}
\arguments{
\item{object}{object of class dare}
}
\description{
Bayesian Information criteria for dare objects.  Lower is better.
}
