% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blinding.R
\name{blinding}
\alias{blinding}
\title{Splitting data to construct overlays that mimic the four sections of
data used to calculate a one-way ANOVA.}
\usage{
blinding(data, variables = NULL)
}
\arguments{
\item{data}{tibble or data frame: dataset containing the variables.}

\item{variables}{A vector of column names or indices to be blinded, with their group order shuffled.
Keep in mind that only groups from nominal variables should be blinded and shuffled.}
}
\description{
Splitting data to construct overlays that mimic the four sections of
data used to calculate a one-way ANOVA.
}
