% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookExercise.R
\name{bookExercise}
\alias{bookExercise}
\title{Open up shiny app that includes interactive exercises that the user can use
to practice.}
\usage{
bookExercise(chapter = 1)
}
\arguments{
\item{chapter}{Integer: Which chapter would you like to see exercises on?}
}
\value{
A shiny app that includes interactive exercises that the user can use
to practice.
}
\description{
Open up shiny app that includes interactive exercises that the user can use
to practice.
}
\examples{
\dontrun{
bookExercise(chapter = 3)
}
}
