% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centToFaren.R
\name{centToFaren}
\alias{centToFaren}
\title{Convert a temperature from Celsius to Fahrenheit}
\usage{
centToFaren(farenTemp)
}
\arguments{
\item{farenTemp}{Temperature, in Fahrenheit}
}
\value{
Temperature, in Celsius
}
\description{
This function converts a temperature from Celsius to Fahrenheit.
}
