% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.tranchor}
\alias{cv.tranchor}
\title{Cross-validating distributional anchor regression models}
\usage{
\method{cv}{tranchor}(
  x,
  epochs,
  xi = NULL,
  folds = NULL,
  fold = 5,
  verbose = TRUE,
  print_training_info = FALSE,
  ...
)
}
\arguments{
\item{x}{Object of class \code{"tranchor"}.}

\item{epochs}{Number of epochs.}

\item{xi}{Optional; can be used to overwrite \code{xi} used when building
\code{x}.}

\item{folds}{Optional vector of length <number of samples> specifying the
folds.}

\item{fold}{If \code{folds} is not specified, folds are computed from the
anchors. See details.}

\item{verbose}{Whether to show a progress bar; defaults to \code{TRUE}.}

\item{print_training_info}{Whether to show the training steps in each fold;
defaults to \code{FALSE}.}

\item{...}{Arguments passed to \code{\link[tranchor]{fit.tranchor}}.}
}
\value{
List of negative log-likelihood contributions in- and out-of-sample,
    \code{folds} and \code{fold}.
}
\description{
Cross-validating distributional anchor regression models
}
