% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dare_gof.R
\name{dare_gof}
\alias{dare_gof}
\title{Goodness of fit Bayesian p-value}
\usage{
dare_gof(object, n_posterior_draws)
}
\arguments{
\item{object}{dare object}

\item{n_posterior_draws}{integer.  Number of posterior draws for the
Bayesian p-value}
}
\value{
a named list giving the Bayesian p-value and the
test statistics for the observed and simulated data.
}
\description{
Use Bayesian p-values to determine if the model fit is appropriate.
}
\details{
Bayesian p-values determine whether data generated from the posterior
predictive distribution (i.e., data generated according to the model,
without having to specify specific parameter values) match the
observed data.  A good fit is indicated by values close to 0.5.  Any
values less than, say, 0.05 or greater than 0.95 indicate a poor fit.
***NOTE: Current implementation is only for the beta-Poisson DR model.
Also note that future.apply package is used here, so to parallelize
(highly recommended), use plan(multicore), or for Windows users use
plan(cluster, workers = \if{html}{\out{<number of cores>}}).  See the future package
for more details.
}
