% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discDist.R
\name{discDist}
\alias{discDist}
\title{borrowed from vcd::distplot, written by Achim Zeileis}
\usage{
discDist(
  x,
  type = c("poisson", "binomial", "nbinomial"),
  size = NULL,
  lambda = NULL,
  conf_int = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{x}{either a vector of counts, a 1-way table of frequencies of counts
or a data frame or matrix with frequencies in the first column and the
corresponding counts in the second column.}

\item{type}{character. String indicating the distribution.}

\item{size}{the size argument for the binomial and negative binomial
distribution. If set to NULL and type is "binomial", then size is taken to
be the maximum count. If set to NULL and type is "nbinomial", then size is
estimated from the data.}

\item{lambda}{parameter of the poisson distribution. If type is "poisson"
and lambda is specified a leveled poissonness plot is produced.}

\item{conf_int}{logical. Should confidence intervals be calculated?}

\item{conf_level}{numeric. Confidence level for confidence intervals.}
}
\value{
Returns a named list containing the parameter estimate(s), the
legend text, the plot title, phat, and a tibble with data used to
construct the plot.
}
\description{
Computes the data for diagnostic distribution plots:
poissonness, binomialness and negative binomialness plots.
}
\examples{
set.seed(1111)
rb <- tibble(k = rbinom(10000, 20, 0.8))
ddOut <- discDist(rb$k, type = "binomial", size = 20)
forPlot <- ddOut$forPlot
ggplot(data = forPlot, aes(x = count, y = metameter)) +
geom_point(color = "red") +
geom_abline(slope = ddOut$estimates$slope, intercept = ddOut$estimates$intercept) +
geom_errorbar(aes(ymin = lower, ymax = upper), width = 0.2)
}
\references{
D. C. Hoaglin (1980), A poissonness plot, The American
Statistican, 34, 146--149.

D. C. Hoaglin & J. W. Tukey (1985), Checking the shape of discrete
distributions. In D. C. Hoaglin, F. Mosteller, J. W. Tukey (eds.),
Exploring Data Tables, Trends and Shapes, chapter 9. John Wiley & Sons,
New York.

M. Friendly (2000), Visualizing Categorical Data. SAS Institute, Cary, NC.
}
