% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fit.dare}
\alias{fit.dare}
\title{Fit distributional anchor regression models}
\usage{
\method{fit}{dare}(
  object,
  epochs = 10,
  early_stopping_metric = "loss",
  callbacks = list(callback_reduce_lr_on_plateau("loss", patience = 30),
    callback_early_stopping("loss", patience = 50)),
  ...
)
}
\arguments{
\item{object}{Object of class \code{"dare"}.}

\item{epochs}{Number of epochs to train the model.}

\item{early_stopping_metric}{Metric monitored for early stopping. Defaults to
\code{"loss"}.}

\item{callbacks}{List of optional callbacks.}

\item{...}{See \code{\link[deepregression]{fit.deepregression}}}
}
\value{
See \code{\link[keras]{fit.keras.engine.training.Model}}
}
\description{
Fit distributional anchor regression models
}
