% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldedLog.R
\name{foldedLog}
\alias{foldedLog}
\title{Calculate a folded log with the Tukey and Haldane-Anscombe corrections}
\usage{
foldedLog(successes, failures)
}
\arguments{
\item{successes}{Integer: numerator frequency.}

\item{failures}{Integer: denominator frequency.}
}
\value{
A 1 x 2 tibble, with columns "Tukey" and "HaldaneAnscombe".
}
\description{
Calculate a folded log with the Tukey and Haldane-Anscombe corrections
}
