% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_overlay.R
\name{gt_overlay}
\alias{gt_overlay}
\title{Summarize ANOVA overlays as a gt_effect_table.}
\usage{
gt_overlay(dareANOVAobj, latex = F)
}
\arguments{
\item{dareANOVAobj}{Object with dare_overlay class}

\item{latex}{Should gt_overlay output LaTeX code?}
}
\value{
a gt_effect_table containing variables and effects from overlay
}
\description{
Summarize ANOVA overlays as a gt_effect_table.
}
\examples{
\dontrun{
# One variable
ANOVA_split(len~dose, ToothGrowth) |> gt_overlay()

# Two variables
ANOVA_split(len~supp*dose, ToothGrowth) |> gt_overlay()
}
}
