% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HQC.dare.R
\name{HQC.dare}
\alias{HQC.dare}
\title{HQC}
\usage{
\method{HQC}{dare}(object)
}
\arguments{
\item{object}{object of class dare}
}
\description{
Hannan-Quinn Information criteria for dare objects.  Lower is better.
}
