% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HQC.R
\name{HQC}
\alias{HQC}
\title{Generic HQC function}
\usage{
HQC(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic function for computing the Hannan-Quinn Information Criterion.
}
