% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{LehmannAN}
\alias{LehmannAN}
\title{Reverse-time proportional hazards anchor regression}
\usage{
LehmannAN(
  formula,
  data,
  anchor,
  xi = 0,
  response_type = get_response_type(data[[all.vars(formula)[1]]]),
  order = get_order(response_type, data[[all.vars(formula)[1]]]),
  addconst_interaction = 0,
  latent_distr = "gumbel",
  monitor_metrics = NULL,
  trafo_options = trafo_control(order_bsp = order, response_type = response_type),
  ...
)
}
\arguments{
\item{formula}{Formula for specifying the regression of response on covariates}

\item{data}{Data.frame or list containing the variables occuring in \code{formula}
and \code{anchor}}

\item{anchor}{Formula for specifying the anchor. The score residuals are
projected onto the columns of the design matrix corresponding to
\code{anchor} and \code{data}}

\item{xi}{Controls strength of causal regularization. \code{xi = 0} is the
unpenalized model.}

\item{response_type}{See \code{\link[deeptrafo]{deeptrafo}}.}

\item{order}{See \code{\link[deeptrafo]{deeptrafo}}.}

\item{addconst_interaction}{See \code{\link[deeptrafo]{deeptrafo}}.}

\item{latent_distr}{See \code{\link[deeptrafo]{deeptrafo}}.}

\item{monitor_metrics}{See \code{\link[deeptrafo]{deeptrafo}}.}

\item{trafo_options}{See \code{\link[deeptrafo]{deeptrafo}}.}

\item{...}{
  Arguments passed on to \code{\link[=tranchor]{tranchor}}
  \describe{
    \item{\code{return_data}}{See \code{\link[deeptrafo]{deeptrafo}}.}
    \item{\code{loss}}{Under development. Please use the default \code{loss = "anchor"}.}
    \item{\code{bw}}{Under development.}
  }}
}
\value{
See \code{\link[tranchor]{tranchor}}
}
\description{
Reverse-time proportional hazards anchor regression
}
