% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.dare}
\alias{logLik.dare}
\title{Log-likelihood for in- and out-of-sample model evaluation}
\usage{
\method{logLik}{dare}(
  object,
  newdata = NULL,
  convert_fun = function(x, ...) -sum(x, ...),
  ...
)
}
\arguments{
\item{object}{Object of class \code{"dare"}}

\item{newdata}{Data.frame or list for evaluating on new data.}

\item{convert_fun}{See \code{\link[deeptrafo]{logLik.deeptrafo}}.}

\item{...}{See \code{\link[deeptrafo]{logLik.deeptrafo}}.}
}
\description{
Log-likelihood for in- and out-of-sample model evaluation
}
