% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dare_overlay.R
\name{plot.dare_overlay}
\alias{plot.dare_overlay}
\title{Summarize ANOVA overlays as a plot.}
\usage{
\method{plot}{dare_overlay}(dareANOVAobj)
}
\arguments{
\item{dareANOVAobj}{Object with dare_overlay class}
}
\value{
a plot containing scatterplot of overlay effects and boxplot of overlay residuals
}
\description{
Summarize ANOVA overlays as a plot.
}
\examples{
\dontrun{
# One variable
ANOVA_split(len~dose, ToothGrowth) |> plot()

# Two variables
ANOVA_split(len~supp*dose, ToothGrowth) |> plot()
}
}
