% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.tranchor}
\alias{residuals.tranchor}
\title{Residuals for in- and out-of-sample model evaluation}
\usage{
\method{residuals}{tranchor}(
  object,
  newdata = NULL,
  convert_fun = function(x, ...) -identity(x),
  ...
)
}
\arguments{
\item{object}{Object of class \code{"tranchor"}}

\item{newdata}{Optional data.frame or list of new data to compute residuals on}

\item{convert_fun}{See \code{\link[deeptrafo]{residuals.deeptrafo}}}

\item{...}{See \code{\link[deeptrafo]{residuals.deeptrafo}}}
}
\value{
Single-column matrix of score residuals
}
\description{
Residuals for in- and out-of-sample model evaluation
}
