% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{searchConjunction}
\alias{searchConjunction}
\title{SearchConjunction dataset}
\format{
\subsection{\code{searchConjunction}}{

A tibble with 39,962 rows and 8 columns:
\describe{
\item{subject}{The subjects in the study}
\item{trial}{Trial ID}
\item{setSize, setSize.f}{The size of the set. Either 3, 6, 12, or 18.}
\item{target}{Whether the target is absent or present}
\item{error}{Whether there was an error or not}
\item{sdt}{Either CR, FA, H, or M}
\item{rTime}{Reaction time}
...
}
}
}
\source{
insert source here (url's can be in < >)
}
\usage{
searchConjunction
}
\description{
Insert description here
}
\keyword{datasets}
