% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sFrac.R
\name{sFrac}
\alias{sFrac}
\title{Calculate a started fraction with the Tukey and Haldane-Anscombe corrections}
\usage{
sFrac(successes, failures, total_count = successes + failures)
}
\arguments{
\item{successes}{Integer: numerator frequency.}

\item{failures}{Integer: denominator frequency.}
}
\value{
A 1 x 2 tibble, with columns "Tukey" and "HaldaneAnscombe".
}
\description{
Calculate a started fraction with the Tukey and Haldane-Anscombe corrections
}
