% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{speedDating}
\alias{speedDating}
\title{Speed Dating dataset}
\format{
\subsection{\code{speedDating}}{

A tibble with 6 rows and 3 columns:
\describe{
\item{goal.f}{A factor, describing the primary goal in participating in the speed dating study.}
\item{n}{The frequency of participants who chose the given primary goal.}
\item{prop}{The proportion of participants who chose the given primary goal.}
}
}
}
\source{
insert source here (url's can be in < >)
}
\usage{
speedDating
}
\description{
Insert description here
}
\keyword{datasets}
