% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dare.R
\name{summary.dare}
\alias{summary.dare}
\title{Summarizing DARE model fits}
\usage{
\method{summary}{dare}(object, CI_level, print = TRUE)
}
\arguments{
\item{object}{object of class "dare"}

\item{CI_level}{the level of credible intervals to be provided}

\item{print}{logical}
}
\description{
summary method for class "dare"
}
