% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tranchor.R
\name{tranchor}
\alias{tranchor}
\title{Distributional anchor regression models}
\usage{
tranchor(
  formula,
  data,
  anchor,
  xi = 0,
  response_type = get_response_type(data[[all.vars(fml)[1]]]),
  order = get_order(response_type, data[[all.vars(fml)[1]]]),
  addconst_interaction = 0,
  latent_distr = "logistic",
  monitor_metrics = NULL,
  trafo_options = trafo_control(order_bsp = order, response_type = response_type),
  return_data = FALSE,
  ...
)
}
\description{
Distributional anchor regression models
}
