% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifSampDist.R
\name{unifSampDist}
\alias{unifSampDist}
\title{The}
\usage{
unifSampDist(n, R, d)
}
\arguments{
\item{n}{Number of observations in the sample}

\item{R}{Number of samples. An integer that is at least 1.}

\item{d}{The minimum and maximum limits for the error}
}
\value{
A dataset
}
\description{
This function
}
