% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA_overlay.R
\name{ANOVA_overlay}
\alias{ANOVA_overlay}
\title{Splitting your data to construct overlays that mimic the four sections of
data used to calculate a one-way ANOVA.}
\usage{
ANOVA_overlay(x, ...)
}
\arguments{
\item{x}{Either your independent variable (as a character), or a formula
in the form dv ~ iv, where dv is your dependent variable, and iv is your
independent variable.}

\item{y}{Character: What is the name of your dependent variable?}

\item{data}{tibble or data frame: Your dataset containing your variables.}

\item{common}{function: which function should represent the common in the
splitting algorithm?}
}
\value{
a list of four tibbles, representing the four overlays: the data,
the common, the condition, and the residuals.
}
\description{
Splitting your data to construct overlays that mimic the four sections of
data used to calculate a one-way ANOVA.
}
\examples{
\dontrun{
# using dplyr to see the data split by condition, assuming a mean common
ToothGrowth \%>\% group_by(supp) \%>\% summarize(mn = mean(len))

# using ANOVA_overlay
ANOVA_overlay("supp", "len", ToothGrowth)
ANOVA_overlay("supp", "len", ToothGrowth, common = median)
ANOVA_overlay("dose", "len", ToothGrowth)

# using ANOVA_overlay with formula
ANOVA_overlay(len~supp, ToothGrowth)

# Showing that the sums of the overlays add up to the data

toothGrowthOverlays <- ANOVA_overlay("supp", "len", ToothGrowth)
SS_data <- sum((toothGrowthOverlays$data)^2)
SS_overlays <- sum((toothGrowthOverlays$common)^2) + sum((toothGrowthOverlays$condition)^2) +
 sum((toothGrowthOverlays$residuals)^2)
all.equal(SS_data,SS_overlays)
}
}
