% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA_overlay.R
\name{ANOVA_overlay}
\alias{ANOVA_overlay}
\title{Splitting your data to construct overlays that mimic the four sections of
data used to calculate a one-way ANOVA.}
\usage{
ANOVA_overlay(data, common = mean, ...)
}
\arguments{
\item{data}{tibble or data frame: Your dataset containing your variables.}

\item{common}{function: which function should represent the common in the
splitting algorithm?}

\item{dv}{Character: What is the name of your dependent variable?}

\item{iv}{Character: What is the name of your independent variable?}

\item{formula}{Formula: A formula in the form dv ~ iv, where dv is your
dependent variable, and iv is your independent variable.}
}
\value{
a list of four tibbles, representing the four overlays: the data,
the common, the condition, and the residuals.
}
\description{
Splitting your data to construct overlays that mimic the four sections of
data used to calculate a one-way ANOVA.
}
\examples{
\dontrun{
# using dplyr to see the data split by condition, assuming a mean common
ToothGrowth \%>\% group_by(supp) \%>\% summarize(mn = mean(len))

# using ANOVA_overlay
ANOVA_overlay(data = ToothGrowth, dv = "len", iv = "supp")
ANOVA_overlay(data = ToothGrowth, dv = "len", iv = "supp", common = median)
ANOVA_overlay(data = ToothGrowth, dv = "len", iv = "dose")

# using ANOVA_overlay with formula
ANOVA_overlay(data = ToothGrowth, formula = len~supp)

# Showing that the sums of the overlays add up to the data

toothGrowthOverlays <- ANOVA_overlay(data = ToothGrowth, dv = "len", iv = "supp")
SS_data <- sum((toothGrowthOverlays$data)^2)
SS_overlays <- sum((toothGrowthOverlays$common)^2) + sum((toothGrowthOverlays$condition)^2) +
 sum((toothGrowthOverlays$residuals)^2)
all.equal(SS_data,SS_overlays)
}
}
