% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dare.R
\name{dare}
\alias{dare}
\title{DARE: Dose Accrual Risk Estimation}
\usage{
dare(
  formula,
  data,
  dose_response = c("beta-poisson", "exponential", "simple_threshold")[1],
  prior_regr_coefs = list(mean = 0, sd = 2.5, autoscale = TRUE),
  prior_regr_intercept = list(mean = 0, sd = 10),
  prior_sigma = list(shape = 2, rate = 2),
  prior_alpha_rate = 1,
  nonlinear_time = FALSE,
  CI_level = 0.95,
  verbose = TRUE,
  optim_args = list(method = "BFGS")
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.  Format should be
\if{html}{\out{<response>}} ~ <fixed effect 1> + <fixed effect 2> + ... + (\if{html}{\out{<time variable>}} | \if{html}{\out{<id>}})}

\item{data}{a data.frame containing the variables in the model.}

\item{dose_response}{The dose-response model to be used.  Either the
beta-poisson, exponential.  The simple threshold model is not finished yet.}

\item{prior_regr_coefs}{a named list giving the hyperparameters of the
regression coefficients, excluding the intercept.  Should include the following:
\itemize{
\item mean - mean of the normal prior
\item sd - standard deviation of the normal prior
\item autoscale - logical used to determine whether the prior sd should be
rescaled according to the standard deviation of the columns of the design matrix
}}

\item{prior_regr_intercept}{named list, same as prior_regr_coefs, except no autoscale argument}

\item{prior_sigma}{named list, giving the shape and rate of the gamma prior on sigma,
the standard deviation of the mean dose distribution}

\item{prior_alpha_rate}{positive value giving the rate of the exponential
prior on alpha (for beta-poisson model only)}

\item{nonlinear_time}{logical.  Set to TRUE if the expected dose should be the rate
times time raised to some power to be estimated. NOTE: this loses the interpretability
of the regression coefficients somewhat.}

\item{CI_level}{level for the credible intervals.}

\item{verbose}{logical. Should progress be displayed?}

\item{optim_args}{named list giving arguments to optim().}
}
\value{
An object of class "dare", which is really a list with the following
named elements:
\itemize{
\item summary - data.frame giving the posterior median, the posterior SD,
the credible intervals (at level CI_level), and the probability of direction.
\item CI_level
\item log_posterior - value of the log posterior evaluated at the MAP.
Useful for information criteria.
\item asymptotic_covariance - Full posterior covariance matrix.
}
}
\description{
Use longitudinal data to estimate risk ratios of accruing dose
}
